requirements.txt
telethon==1.36.0
anthropic==0.25.0
python-dotenv==1.0.0
apscheduler==3.10.4
aiohttp==3.9.5
fastapi==0.111.0
uvicorn==0.30.1
supabase==2.4.6
PyGithub==2.3.0
python-multipart==0.0.9
google-api-python-client==2.130.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
